+++ src/main/java/org/zkollonay/stockmate/service/ExchangeRateService.java
@@ -2,11 +2,13 @@ package org.zkollonay.stockmate.service;
 
 import org.zkollonay.stockmate.DTO.ExchangeRateDTO;
 
+import java.math.BigDecimal;
+
 public interface ExchangeRateService {
 
-  double fetchExchangeRate(String fromCurrency, String toCurrency);
+  BigDecimal fetchExchangeRate(String fromCurrency, String toCurrency);
 
   ExchangeRateDTO getExchangeRateDetails(String fromCurrency, String toCurrency);
 
-  ExchangeRateDTO convertToCurrency(double amount, String fromCurrency, String toCurrency);
+  ExchangeRateDTO convertToCurrency(BigDecimal amount, String fromCurrency, String toCurrency);
 }
